refactor HtmlSanitizerConfiguration to match ABP's extendable configuration structure
