+++ src/Abp.HtmlSanitizer/HtmlSanitizer/AbpHtmlSanitizerExtensions.cs
@@ -1,11 +1,12 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 
-namespace Abp.HtmlSanitizer;
-
-public static class AbpHtmlSanitizerExtensions
+namespace Abp.HtmlSanitizer
 {
-    public static void AddAbpHtmlSanitizer(this MvcOptions options)
+    public static class AbpHtmlSanitizerExtensions
     {
-        options.Filters.AddService(typeof(AbpHtmlSanitizerActionFilter));
+        public static void AddAbpHtmlSanitizer(this MvcOptions options)
+        {
+            options.Filters.AddService(typeof(AbpHtmlSanitizerActionFilter));
+        }
     }
 }
