+++ src/Abp.HtmlSanitizer/HtmlSanitizer/AbpHtmlSanitizerModule.cs
@@ -1,17 +1,23 @@
 ﻿using System.Reflection;
 using Abp.Dependency;
+using Abp.HtmlSanitizer.Configuration;
 using Abp.Modules;
 using Ganss.Xss;
 
-namespace Abp.HtmlSanitizer;
-
-[DependsOn(typeof(AbpKernelModule))]    
-public class AbpHtmlSanitizerModule : AbpModule
+namespace Abp.HtmlSanitizer
 {
-    public override void Initialize()
+    [DependsOn(typeof(AbpKernelModule))]    
+    public class AbpHtmlSanitizerModule : AbpModule
     {
-        IocManager.RegisterAssemblyByConvention(Assembly.GetExecutingAssembly());
-        
-        IocManager.Register<IHtmlSanitizer, Ganss.Xss.HtmlSanitizer>(DependencyLifeStyle.Transient);
+        public override void PreInitialize()
+        {
+            IocManager.Register<IHtmlSanitizerConfiguration, HtmlSanitizerConfiguration>();
+        }
+
+        public override void Initialize()
+        {
+            IocManager.RegisterAssemblyByConvention(Assembly.GetExecutingAssembly());
+            IocManager.Register<IHtmlSanitizer, Ganss.Xss.HtmlSanitizer>(DependencyLifeStyle.Transient);
+        }
     }
 }
