+++ src/Abp.HtmlSanitizer/HtmlSanitizer/DisableHtmlSanitizerAttribute.cs
@@ -1,8 +1,9 @@
 ﻿using System;
 
-namespace Abp.HtmlSanitizer;
-
-[AttributeUsage(AttributeTargets.Class | AttributeTargets.Method | AttributeTargets.Property)]
-public class DisableHtmlSanitizerAttribute : Attribute{
+namespace Abp.HtmlSanitizer
+{
+    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Method | AttributeTargets.Property)]
+    public class DisableHtmlSanitizerAttribute : Attribute{
     
+    }
 }
