+++ src/Abp.HtmlSanitizer/HtmlSanitizer/HtmlSanitizerAttribute.cs
@@ -1,9 +1,10 @@
 ﻿using System;
 
-namespace Abp.HtmlSanitizer;
-
-[AttributeUsage(AttributeTargets.Class | AttributeTargets.Method | AttributeTargets.Property)]
-public class HtmlSanitizerAttribute : Attribute
+namespace Abp.HtmlSanitizer
 {
+    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Method | AttributeTargets.Property)]
+    public class HtmlSanitizerAttribute : Attribute
+    {
     
+    }
 }
